Moved `crash_handler` to separate folder
